the core defines a gate_id to be a uint8_t, let's use that type everywhere else
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@17118 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ nmad/protocols/proto_mpi/include/mpi_nmad_private.h
@@ -152,7 +152,7 @@ typedef struct mpir_request_s {
   /** communication mode to be used when exchanging data */
   MPI_Communication_Mode communication_mode;
   /** gate of the destination or the source node */
-  uint16_t gate_id;
+  uint8_t gate_id;
   /** number of elements to be exchanged */
   int count;
   /** pointer to the data to be exchanged */
@@ -242,22 +242,22 @@ int mpir_internal_exit(void);
 /**
  * Gets the incoming gate id for the given node
  */
-uint16_t mpir_get_in_gate_id(int node);
+uint8_t mpir_get_in_gate_id(int node);
 
 /**
  * Gets the outgoing gate id for the given node
  */
-uint16_t mpir_get_out_gate_id(int node);
+uint8_t mpir_get_out_gate_id(int node);
 
 /**
  * Gets the node associated to the given incoming gate id
  */
-int mpir_get_in_dest(uint16_t gate);
+int mpir_get_in_dest(uint8_t gate);
 
 /**
  * Gets the node associated to the given outgoing gate id
  */
-int mpir_get_out_dest(uint16_t gate);
+int mpir_get_out_dest(uint8_t gate);
 
 /* Send/recv/status functions */
